requirements.txt
pip
setuptools
os
fernet
cv2
pygame
GitPython
time
schedule
platform
cryptography
pynput